work 1 march

+++ sim_some_params_frozen/percolation.py
@@ -7,6 +7,7 @@ import pandas as pd
 from scipy.optimize import curve_fit
 from numpy import linspace
 from collections import Counter
+import os
 import sys
 sys.path.append('..')
 from global_functions import *
@@ -97,37 +98,25 @@ def componentsHistogram(N, interac_r, exclusion_r, arena_r, excludeGiantComp = T
     fig.savefig(f'com_sizes_histogram_N_{N}_ar_{arena_r}_ir_{interac_r}_quenched_zoom.png')
     plt.close(fig)
     
-def plotComSizes_difN(Ns, arena_rs, interac_rs, exclusion_r, prob=False, push=False, excludeGiantComp = False, dataToFile = False):
+def plotComSizes_difN(Ns, arena_rs, interac_rs, exclusion_r, logBins_l, push=False, excludeGiantComp = False, dataToFile = False):
     pushLabel = 'push' if push else 'nopush'
     gcLabel = 'excludedGC' if excludeGiantComp else ''
     fig, ax = plt.subplots()
-    ylabel = 'P(s)' if prob else 'N(s)'
-    ax.set(xlabel='s', ylabel=ylabel, xscale='log', yscale='log')
-    for N, ar, ir in zip(Ns, arena_rs, interac_rs):
+    ax.set(xlabel='s', ylabel='P(s)', xscale='log', yscale='log')
+    for N, ar, ir, logBins in zip(Ns, arena_rs, interac_rs, logBins_l):
         com_sizes = getCommunitySizes(N, exclusion_r, ir, ar, push, excludeGiantComp, exclusionVersion=0)
-        # fer el conteig de N(s)
-        com_sizes_counter = Counter(com_sizes)
-        coms = list(com_sizes_counter.keys())
-        counts = list(com_sizes_counter.values())
-        probs = [count/len(com_sizes) for count in list(com_sizes_counter.values())]
-        # logarithmic binning:
-        box_edges = np.geomspace(10, max(coms),10)
-        
-        #irs = [ir]*len(coms)
-        com_counts_df = pd.DataFrame({'coms':coms, 'counts':counts, 'probs':probs})
-        com_counts_df = com_counts_df.sort_values(by='coms', ignore_index=True)
+        binLims, binCenters = binsForHist1D_log((min(com_sizes), max(com_sizes)), logBins)
+        binCenters, hist, stdHist = hist1D(com_sizes, binLims, binCenters, isPDF=True)
+        com_counts_df = pd.DataFrame({'coms':binCenters, 'prob':hist, 'stdprob':stdHist})
         if dataToFile:
             filename = f'comSizesCounts_N_{N}_ar_{ar}_ir_{ir}_er_{exclusion_r}_{pushLabel}_{gcLabel}.csv'
             com_counts_df.to_csv(filename, index=False)
         # plot:
-        if prob:
-            ax.plot(com_counts_df['coms'], com_counts_df['probs'], label=rf'{N}, $r_i^{{*}} = {ir}$', marker='.', ls='None')
-        else:
-            ax.plot(com_counts_df['coms'], com_counts_df['counts'], label=rf'{N}, $r_i^{{*}} = {ir}$', marker='.', ls='None')
+        ax.plot(com_counts_df['coms'], com_counts_df['prob'], label=rf'{N}, $r_i^{{*}} = {ir}$', marker='.', ls='None')
     fig.text(0.25, 0.97, f'$excludeGiantComp = {excludeGiantComp}, push = {push}$')
     fig.legend(title='N', loc=(0.7,0.7))
     fig.tight_layout()
-    figName = 'comSizesProbs' if prob else 'comSizesCounts'
+    figName = 'comSizesProbs'
     for N in Ns:
         figName += f'_{N}'
     figName += f'_er_{exclusion_r}_{pushLabel}_{gcLabel}.png'
@@ -137,13 +126,16 @@ def plotComSizes_difN(Ns, arena_rs, interac_rs, exclusion_r, prob=False, push=Fa
 def getDegreeDistr(N, arena_r, interac_r, exclusion_r, push=False, toFile=True):
     pushLabel = 'push' if push else 'nopush'
     filename = f'other_res_files/degreeDistr_N_{N}_ar_{arena_r}_ir_{interac_r}_er_{exclusion_r}_{pushLabel}.csv'
-    degrees = getBotDegrees(N, exclusion_r, interac_r, arena_r, push=push)
-    binCenters = np.linspace(min(degrees), max(degrees), max(degrees)-min(degrees)+1)
-    binLims = np.linspace(min(degrees), max(degrees)+1, max(degrees)+1-min(degrees)+1) - 0.5
-    binCenters, prob, dprob = hist1D(degrees, binLims, binCenters, isPDF = True)
-    df = pd.DataFrame({'binCenters':binCenters, 'prob':prob, 'dprob':dprob})
-    if toFile:
-        df.to_csv(filename, index=False)
+    if os.path.exists(filename):
+        df = pd.read_csv(filename)
+    else:
+        degrees = getBotDegrees(N, exclusion_r, interac_r, arena_r, push=push)
+        binCenters = np.linspace(min(degrees), max(degrees), max(degrees)-min(degrees)+1)
+        binLims = np.linspace(min(degrees), max(degrees)+1, max(degrees)+1-min(degrees)+1) - 0.5
+        binCenters, prob, dprob = hist1D(degrees, binLims, binCenters, isPDF = True)
+        df = pd.DataFrame({'binCenters':binCenters, 'prob':prob, 'dprob':dprob})
+        if toFile:
+            df.to_csv(filename, index=False)
     return df
     
 def plotDegreeDistr_manyir(N, arena_r, irs, exclusion_r, push=False, poisson=False):
@@ -156,6 +148,9 @@ def plotDegreeDistr_manyir(N, arena_r, irs, exclusion_r, push=False, poisson=Fal
             avgdeg = np.mean(degrees)
             expo = np.exp(-avgdeg)
             df['poisson'] = expo*(avgdeg**df['binCenters'])/(factorial(df['binCenters']))
+            pushLabel = 'push' if push else 'nopush'
+            filename = f'other_res_files/degreeDistr_N_{N}_ar_{arena_r}_ir_{ir}_er_{exclusion_r}_{pushLabel}.csv'
+            df.to_csv(filename, index=False)
             ax.plot(df['binCenters'], df['poisson'], lw=0.8, color = line.get_color(), alpha=0.5)
     ax.set(xlabel='Degree, $k$', ylabel='$P(k)$')
     fig.legend(loc=(0.7, 0.6), fontsize=9)
@@ -184,10 +179,9 @@ def main():
     #componentsHistogram(492, 6.0, 1.5, 75.0)
     ####
     # a dia 21 de febrer 2023 ho estic fent amb 1000 configs N35, 250 configs N492
-    # plotComSizes_difN([35,492], [20.0, 75.0], [6.5, 6.4], 1.5, prob=True, push=False, excludeGiantComp = False, dataToFile=True)
-    # plotComSizes_difN([35,492], [20.0, 75.0], [6.5, 6.4], 1.5, prob=True, push=False, excludeGiantComp = True, dataToFile=True)
-    plotDegreeDistr_manyir(35, 20.0, [5.0, 6.5, 8.0, 9.0], 1.5, poisson=True)
-    plotDegreeDistr_manyir(492, 75.0, [4.9, 6.4, 7.6, 9.1], 1.5, poisson=True)
+    plotComSizes_difN([35,492], [20.0, 75.0], [6.5, 6.4], 1.5, logBins_l = [10, 40], push=False, excludeGiantComp = True, dataToFile=True)
+    # plotDegreeDistr_manyir(35, 20.0, [5.0, 6.5, 8.0, 9.0], 1.5, poisson=True)
+    # plotDegreeDistr_manyir(492, 75.0, [4.9, 6.4, 7.6, 9.1], 1.5, poisson=True)
